CP navigation / Simple AI created and moving -Brendon
Both AI and player will automatically navigate the track.
Looks quite messy at the moment, have tried to comment the things that will be removed/ moved once the players car is instantiated properly.
Getting the AI cars and player car to instantiate correctly shouldn't be that big of a deal.